Split test/content.js into several files.
